Add 'password' field to user profile
The password field was added to the user profile input data in order to improve data security. As a result, the createUser function in the UserCubit, UserProfile, UserRepository, UserEntity and User model were all updated to include the password parameter. Additionally, text for password requirement was added to en-US translations. This commit contributes to better user authentication management.