Add Sidebar and fix incorrect Eithers in Auth
This commit introduces a new Sidebar widget and makes changes to the 'auth' domain. It also includes miscellaneous updates such as dependency versions. In the 'auth' domain, the incorrect type position in the Either used for sign-in related functions was fixed. This improves the error handling as now the failure case is correctly represented by the left side of Either. Moreover, within Sign Up data, null checks were removed as data class initialization already takes care of it. Some initial setup for intercom is also included. Lastly, the versions for dependencies were updated for compatibility and taking advantage of new features/fixes.