Add white app logo, update default avatar and bump version
A white version of the mobile app logo has been added to the assets. This 'imgAppLogoWhite' image is now used as the default avatar in user profiles, replacing 'imgAppstore'. This change provides a more consistent branding experience across the app. Moreover, the app version has been bumped to '1.0.2+1' to reflect these updates.